+++ include/modules/Blossom.h
@@ -7,10 +7,10 @@
 
 namespace dhe {
 
-static auto constexpr bounceRange = Range{1.F, 17.F};
-static auto constexpr spinKnobTaperCurvature = -0.8F;
-static auto constexpr spinKnobTaper = taper::FixedSTaper{spinKnobTaperCurvature};
-static auto constexpr spinRange = Range{-10.F, 10.F};
+static auto constexpr ratioRange = Range{1.F, 17.F};
+static auto constexpr speedKnobTaperCurvature = -0.8F;
+static auto constexpr speedKnobTaper = taper::FixedSTaper{speedKnobTaperCurvature};
+static auto constexpr speedRange = Range{-10.F, 10.F};
 
 class Blossom : public rack::engine::Module {
 public:
@@ -18,47 +18,37 @@ public:
   void process(const ProcessArgs &args) override;
 
   enum ParameterIds {
-    SpinKnob,
-    SpinAvKNob,
-    BounceRatioKnob,
-    BounceRatioAvKnob,
-    BounceRatioModeSwitch,
-    BounceDepthKnob,
-    BounceDepthAvKnob,
-    BouncePhaseOffsetKnob,
-    BouncePhaseOffsetAvKnob,
+    SpeedKnob,
+    SpeedAvKNob,
+    RatioKnob,
+    RatioAvKnob,
+    FreeRatioSwitch,
+    DepthKnob,
+    DepthAvKnob,
+    PhaseOffsetKnob,
+    PhaseOffsetAvKnob,
     XGainKnob,
     YGainKnob,
     XRangeSwitch,
     YRangeSwitch,
     ParameterCount
   };
-  enum InputIds {
-    SpinCvInput,
-    BounceRatioCvInput,
-    BounceDepthCvInput,
-    BouncePhaseCvInput,
-    XGainCvInput,
-    YGainCvInput,
-    InputCount
-  };
+  enum InputIds { SpeedCvInput, RatioCvInput, DepthCvInput, PhaseCvInput, XGainCvInput, YGainCvInput, InputCount };
   enum OutputIds { XOutput, YOutput, OutputCount };
 
 private:
-  inline auto bounce() const -> float {
-    return scaledRotation(this, BounceRatioKnob, BounceRatioCvInput, BounceRatioAvKnob, bounceRange);
-  }
+  inline auto ratio() const -> float { return scaledRotation(this, RatioKnob, RatioCvInput, RatioAvKnob, ratioRange); }
 
-  inline auto bounceIsFree() const -> bool { return switchPosition(this, BounceRatioModeSwitch) == 1; }
+  inline auto ratioIsFree() const -> bool { return switchPosition(this, FreeRatioSwitch) == 1; }
 
-  inline auto depth() const -> float { return rotation(this, BounceDepthKnob, BounceDepthCvInput, BounceDepthAvKnob); }
+  inline auto depth() const -> float { return rotation(this, DepthKnob, DepthCvInput, DepthAvKnob); }
 
   inline auto phase() const -> float {
-    return scaledRotation(this, BouncePhaseOffsetKnob, BouncePhaseCvInput, BouncePhaseOffsetAvKnob, {-0.5, 0.5});
+    return scaledRotation(this, PhaseOffsetKnob, PhaseCvInput, PhaseOffsetAvKnob, {-0.5, 0.5});
   }
 
-  inline auto spin() const -> float {
-    return taperedAndScaledRotation(this, SpinKnob, SpinCvInput, SpinAvKNob, spinKnobTaper, spinRange);
+  inline auto speed() const -> float {
+    return taperedAndScaledRotation(this, SpeedKnob, SpeedCvInput, SpeedAvKNob, speedKnobTaper, speedRange);
   }
 
   inline auto xGain() const -> float { return scaledRotation(this, XGainKnob, XGainCvInput, gainRange); }