+++ src/panels/XycloidPanel.cpp
@@ -3,36 +3,40 @@
 namespace dhe {
 
 XycloidPanel::XycloidPanel(Xycloid *module) : Panel{module, hp} {
-  auto widgetRightEdge = width();
+  auto const widgetRightEdge = width();
 
-  auto column1 = widgetRightEdge / 7.F;
-  auto column4 = widgetRightEdge - column1;
-  auto column2 = (column4 - column1) / 3.F + column1;
-  auto column3 = widgetRightEdge - column2;
+  auto const column1 = widgetRightEdge / 7.F;
+  auto const column4 = widgetRightEdge - column1;
+  auto const column2 = (column4 - column1) / 3.F + column1;
+  auto const column3 = widgetRightEdge - column2;
 
-  auto y = 30.F;
-  auto dy = 22.F;
+  auto y = 25.F;
+  auto dy = 18.5F;
 
-  input(column1, y, Xycloid::WobbleRatioCvInput);
-  knob<TinyKnob>(column2, y, Xycloid::WobbleRatioAvKnob);
-  knob<LargeKnob>(column3, y, Xycloid::WobbleRatioKnob);
-  toggle<2>(column4, y, Xycloid::WobbleRatioModeSwitch);
-
-  y += dy;
-  input(column1, y, Xycloid::WobbleDepthCvInput);
-  knob<TinyKnob>(column2, y, Xycloid::WobbleDepthAvKnob);
-  knob<LargeKnob>(column3, y, Xycloid::WobbleDepthKnob);
-  toggle<3>(column4, y, Xycloid::WobbleDirectionSwitch);
+  input(column1, y, Xycloid::SpeedCvInput);
+  knob<TinyKnob>(column2, y, Xycloid::SpeedAvKnob);
+  knob<LargeKnob>(column3, y, Xycloid::SpeedKnob);
 
   y += dy;
-  input(column1, y, Xycloid::ThrobSpeedCvInput);
-  knob<TinyKnob>(column2, y, Xycloid::ThrobSpeedAvKnob);
-  knob<LargeKnob>(column3, y, Xycloid::ThrobSpeedKnob);
-  knob<SmallKnob>(column4, y, Xycloid::WobblePhaseOffsetKnob);
+  input(column1, y, Xycloid::RatioCvInput);
+  knob<TinyKnob>(column2, y, Xycloid::RatioAvKnob);
+  knob<LargeKnob>(column3, y, Xycloid::RatioKnob);
+  toggle<2>(column4, y, Xycloid::FreeRatioSwitch);
+
+  y += dy;
+  input(column1, y, Xycloid::DepthCvInput);
+  knob<TinyKnob>(column2, y, Xycloid::DepthAvKnob);
+  knob<LargeKnob>(column3, y, Xycloid::DepthKnob);
+  toggle<3>(column4, y, Xycloid::DirectionSwitch);
+
+  y += dy;
+  input(column1, y, Xycloid::PhaseCvInput);
+  knob<TinyKnob>(column2, y, Xycloid::PhaseOffsetAvKnob);
+  knob<LargeKnob>(column3, y, Xycloid::PhaseOffsetKnob);
 
   y = 82.F;
   dy = 15.F;
-  const auto outputPortOffset = 1.25F;
+  auto const outputPortOffset = 1.25F;
 
   y += dy;
   input(column1, y, Xycloid::XGainCvInput);