+++ CHANGES.md
@@ -5,6 +5,9 @@
 - All modules
   - Added parameter tooltips.
   - Added the ability to specify parameter values via keyboard.
+- Xycloid
+  - Added phase CV with attenuverter.
+  - Reordered controls to match Blossom.
 - Ranger
   - Renamed and reoriented the limit controls to better convey their effects on the **LEVEL** knob.
 - Swave